requirements.txt
altair==5.4.1
annotated-types==0.7.0
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.9
google-api-core==2.19.2
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.1
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
httplib2==0.22.0
idna==3.10
ja-sentence-segmenter==0.0.2
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
narwhals==1.8.1
numpy==2.1.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.1
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.1
pydantic_core==2.23.3
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
rsa==4.9
six==1.16.0
smmap==5.0.1
streamlit==1.38.0
streamlit-draggable-list==0.0.1
streamlit-sortables==0.2.0
tenacity==8.5.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
watchdog==4.0.2
